refactor:move filtering and tab panel to own component
